+++ src/AttributeEngine/AttributeDoesNotExistException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenDialogAi\Core\Attribute;
+namespace OpenDialogAi\AttributeEngine;
 
 /**
  * Should be thrown when trying to get an attribute from an attribute bag, but one has not been set
