+++ src/AttributeEngine/IntAttribute.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace OpenDialogAi\Core\Attribute;
-
-use Illuminate\Support\Facades\Log;
+namespace OpenDialogAi\AttributeEngine;
 
 /**
  * Int implementation of Attribute.
