+++ src/AttributeEngine/Tests/SecondAbstractAttributeCollection.php
@@ -1,23 +1,25 @@
 <?php
 
-namespace OpenDialogAi\Core\Attribute\Tests;
+namespace OpenDialogAi\AttributeEngine\Tests;
 
-use OpenDialogAi\Core\Attribute\ArrayAttribute;
-use OpenDialogAi\Core\Attribute\Composite\AbstractAttributeCollection;
-use OpenDialogAi\Core\Attribute\IntAttribute;
-use OpenDialogAi\Core\Attribute\Util;
+use OpenDialogAi\AttributeEngine\ArrayAttribute;
+use OpenDialogAi\AttributeEngine\Composite\AbstractAttributeCollection;
+use OpenDialogAi\AttributeEngine\IntAttribute;
+use OpenDialogAi\AttributeEngine\Util;
 
 /**
  * A composite attribute collection containing other attribute types.
+ * This specific composite attribute will also container another composite attribute.
  *
  * createFromInput()
  * @return
  * [
  *  total = IntAttribute,
- *  results = ArrayAttribute
+ *  results = ArrayAttribute,
+ *  test = ExampleAbstractCompositeAttribute
  * ]
  */
-class ExampleAbstractAttributeCollection extends AbstractAttributeCollection
+class SecondAbstractAttributeCollection extends AbstractAttributeCollection
 {
     const EXAMPLE_TYPE = 'api';
     const EXAMPLE_TYPE_ARRAY = 'array';
@@ -46,7 +48,8 @@ class ExampleAbstractAttributeCollection extends AbstractAttributeCollection
      * @return array
      * [
      *  total = IntAttribute,
-     *  results = ArrayAttribute
+     *  results = ArrayAttribute,
+     *  test = ExampleAbstractCompositeAttribute
      * ]
      */
     public function createFromInput($input, $type): array
@@ -56,6 +59,13 @@ class ExampleAbstractAttributeCollection extends AbstractAttributeCollection
         if ($type === self::EXAMPLE_TYPE_ARRAY) {
             $attributes[] = new IntAttribute('total', count($input));
             $attributes[] = new ArrayAttribute('results', $input);
+            $attributes[] = new ExampleAbstractCompositeAttribute(
+                'test',
+                new ExampleAbstractAttributeCollection(
+                    [1 => 'first', 2 => 'second', 3 => 'third'],
+                    ExampleAbstractAttributeCollection::EXAMPLE_TYPE_ARRAY
+                )
+            );
         }
 
         return $attributes;
