+++ src/AttributeEngine/Util.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenDialogAi\Core\Attribute;
+namespace OpenDialogAi\AttributeEngine;
 
 /**
  * Collection of Util/Helper functions.
